src/requirements.txt
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.8
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==25.4.0
beartype==0.22.9
beautifulsoup4==4.14.3
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
contourpy==1.3.3
coolname==4.1.0
cronsim==2.7
cryptography==46.0.5
cryptography-fernet==0.1.0
curl_cffi==0.13.0
cycler==0.12.1
cyclopts==4.10.0
dateparser==1.3.0
docker==7.1.0
docstring_parser==0.17.0
docutils==0.22.4
exceptiongroup==1.3.1
fakeredis==2.34.1
fastapi==0.135.1
fonttools==4.62.1
frozendict==2.4.7
fsspec==2026.2.0
graphviz==0.21
greenlet==3.3.2
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llvmlite==0.44.0
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
multitasking==0.0.12
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
opentelemetry-api==1.40.0
orjson==3.11.7
packaging==26.0
pandas==3.0.1
pandas-ta==0.4.71b0
pathspec==1.0.4
peewee==4.0.1
pendulum==3.2.0
pillow==12.2.0
platformdirs==4.9.2
pluggy==1.6.0
prefect==3.6.22
prometheus_client==0.24.1
protobuf==7.34.0
py-key-value-aio==0.4.4
pyarrow==23.0.1
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydocket==0.18.2
Pygments==2.19.2
PyMySQL==1.1.2
pyodbc==5.3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-json-logger==4.0.0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
readchar==4.2.1
redis==7.3.0
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rich-rst==1.3.2
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
seaborn==0.13.2
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.48
starlette==0.52.1
text-unidecode==1.3
toml==0.10.2
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uncalled-for==0.2.0
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
yfinance==1.2.0
zipp==3.23.0
